Disable localStorage caching, add low-freq date state

+++ scripts/push-update.ps1
@@ -1,4 +1,8 @@
-rm -recurse build/*, serverBuild/*
-npm run build
+param ([Parameter()][switch]$noBuild)
+
+if (-not $noBuild) {
+	rm -recurse build/*, serverBuild/*
+	npm run build
+}
 scp -r config.json package.json package-lock.json scripts/ build/ serverBuild/ steven@192.168.1.52:/home/steven/home-hub
 ssh steven@192.168.1.52 'chmod +x /home/steven/home-hub/scripts/*.sh && /home/steven/home-hub/scripts/process-update.sh'
